% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.full}
\alias{df.full}
\title{random data for testing the functions
data frame with no missing values}
\format{An object of class \code{data.frame} with 100 rows and 51 columns.}
\usage{
df.full
}
\description{
random data for testing the functions
data frame with no missing values
}
\keyword{datasets}
