% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.na}
\alias{df.na}
\title{data frame with missing values}
\format{An object of class \code{data.frame} with 100 rows and 51 columns.}
\usage{
df.na
}
\description{
data frame with missing values
}
\keyword{datasets}
