% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg.weights}
\alias{reg.weights}
\title{Regression weights for SGRQ items}
\format{A data frame with 51 variables (id, sgrq.1 -- sgrq.17) and six rows}
\usage{
reg.weights
}
\description{
A data frame containing the regression weights for SGRQ items
}
\keyword{datasets}
