% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{scoring_sgrq}
\alias{scoring_sgrq}
\title{Severe Respiratory Insufficiency Questionnaire (SRI)}
\usage{
scoring_sgrq(X, id = "")
}
\arguments{
\item{X}{a \code{\link{matrix}} or \code{\link{data.frame}} of 50
columns, containing the questionnaire items. In order from left to right:
sgrq.1 -- sgrq.50.}
}
\description{
SGRQ scoring
}
