% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgrqr.R
\docType{package}
\name{sgrqr}
\alias{sgrqr}
\alias{sgrqr-package}
\title{sgrqr: Scoring the St. George Respiratory Questionnaire (SGRQ)}
\description{
This package can be used to score the St. George Respiratory Questionnaire (SGRQ).
}
\section{Functions}{


`srir` currently provides two functions:
 \code{\link{score_sgrq}} for scoring the SGRQ and
 \code{\link{sum.n}} for defining the minimum number of valid items when calculating summary scores.
}

\references{
http://www.healthstatus.sgul.ac.uk/sgrq
}
